Renaming maybe package to property
