+++ src/main/java/be/jdevelopment/tools/validation/property/MonadOfProperties.java
@@ -1,4 +1,4 @@
-package be.jdevelopment.tools.validation.maybe;
+package be.jdevelopment.tools.validation.property;
 
 public interface MonadOfProperties {
