+++ src/main/java/be/jdevelopment/tools/validation/step/ValidationProcess.java
@@ -3,8 +3,8 @@ package be.jdevelopment.tools.validation.step;
 import be.jdevelopment.tools.validation.ObjectProvider;
 import be.jdevelopment.tools.validation.PropertyToken;
 import be.jdevelopment.tools.validation.error.MonadFactory;
-import be.jdevelopment.tools.validation.maybe.Property;
-import be.jdevelopment.tools.validation.maybe.MonadOfProperties;
+import be.jdevelopment.tools.validation.property.Property;
+import be.jdevelopment.tools.validation.property.MonadOfProperties;
 
 import java.util.*;
