+++ src/test/java/be/jdevelopment/tools/validation/complex/PersonFactory.java
@@ -1,8 +1,8 @@
 package be.jdevelopment.tools.validation.complex;
 
 import be.jdevelopment.tools.validation.ObjectProvider;
-import be.jdevelopment.tools.validation.maybe.Property;
-import be.jdevelopment.tools.validation.maybe.MonadOfProperties;
+import be.jdevelopment.tools.validation.property.Property;
+import be.jdevelopment.tools.validation.property.MonadOfProperties;
 import be.jdevelopment.tools.validation.step.ValidationProcess;
 
 import java.util.*;
